requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
google-generativeai==0.3.2
pydantic==2.5.0
python-dotenv==1.0.0
streamlit==1.28.1
requests==2.31.0
pytz==2023.3
